pyproject.toml
[project]
name = "autoagent-bedrock"
version = "0.1.0"
description = "Bedrock template for always-on Claude research agents"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]
